extracted validate_username function into a separate file
