+++ firstvoices/backend/tests/test_utils.py
@@ -3,7 +3,10 @@ from itertools import product
 import pytest
 from rest_framework import serializers
 
-from backend.serializers.utils.csv_utils import check_required_headers, validate_headers
+from backend.serializers.utils.import_job_utils import (
+    check_required_headers,
+    validate_headers,
+)
 from backend.utils.character_utils import ArbSorter, CustomSorter, nfc
